build calculation engine

+++ backend/services/service_questions.py
@@ -52,7 +52,8 @@ def question_temp_save(req):
     email = get_jwt_identity()
     pack = {
       "email": email,
-      "metadata": req
+      "metadata": req,
+      "count": 0
     }
     db_col.insert_one(pack)
     return make_response(json.dumps({'message': 'success saved'}), 200)
@@ -64,6 +65,10 @@ def question_temp_load(req):
     email = get_jwt_identity()
     pack = db['questions_cache'].find_one({'email': email})
     if pack:
+      if pack['count'] < 1:
+        pack['count'] += 1
+        db['questions_cache'].update_one({'email': email}, {'$set': pack})
+        return make_response(json.dumps({'metadata': pack['metadata']}), 200)
       db['questions_cache'].delete_one({'email': email})
       return make_response(json.dumps({ email: pack['metadata']}), 200)
     else:
@@ -72,4 +77,8 @@ def question_temp_load(req):
     return make_response(json.dumps({'message': 'Server Error'}), 500)
 
 def question_answer(req):
+  ans_pack = req
+  if ans_pack:
+    
   return "success"
+
